pyproject.toml
[project]
name = "aiput"
version = "1.0.0"
description = "通过手机端语音输入实现电脑端远程输入的工具"
authors = [
    {name = "AIPut"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["remote-input", "voice-typing", "wireless-keyboard"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Environment :: X11 Applications",
    "Environment :: Wayland",
]

dependencies = [
    "flask>=3.0.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "qrcode>=7.4.2",
    "pillow>=10.0.0",
    "pystray>=0.19.0",
    "aiohttp>=3.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/newbe36524/AIPut"
Repository = "https://github.com/newbe36524/AIPut"
Issues = "https://github.com/newbe36524/AIPut/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["."]
